agregadas condiciones para evaluar si existen o no las tablas a crear y/o los campos a agregar en migraciones

+++ database/migrations/2019_02_25_015340_create_profiles_table.php
@@ -13,20 +13,22 @@ class CreateProfilesTable extends Migration
      */
     public function up()
     {
-        Schema::create('profiles', function (Blueprint $table) {
-            $table->increments('id');
-            $table->string('first_name')->comment('Nombres');
-            $table->string('last_name')->nullable()->comment('Apellidos');
-            $table->integer('image_id')->nullable()->unsigned()
-                  ->comment('Identificador asociado a la imagen de perfil');
-            $table->foreign('image_id')->references('id')->on('images')->onUpdate('cascade');
-            $table->integer('user_id')->unsigned()->nullable()
-                  ->comment('Identificador del usuario');
-            $table->foreign('user_id')->references('id')
-                  ->on('users')->onDelete('restrict')->onUpdate('cascade');
-            $table->timestamps();
-            $table->softDeletes()->comment('Fecha y hora en la que el registro fue eliminado');
-        });
+        if (!Schema::hasTable('profiles')) {
+            Schema::create('profiles', function (Blueprint $table) {
+                $table->increments('id');
+                $table->string('first_name')->comment('Nombres');
+                $table->string('last_name')->nullable()->comment('Apellidos');
+                $table->integer('image_id')->nullable()->unsigned()
+                      ->comment('Identificador asociado a la imagen de perfil');
+                $table->foreign('image_id')->references('id')->on('images')->onUpdate('cascade');
+                $table->integer('user_id')->unsigned()->nullable()
+                      ->comment('Identificador del usuario');
+                $table->foreign('user_id')->references('id')
+                      ->on('users')->onDelete('restrict')->onUpdate('cascade');
+                $table->timestamps();
+                $table->softDeletes()->comment('Fecha y hora en la que el registro fue eliminado');
+            });
+        }
     }
 
     /**
